requirements.txt
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests
openai
supabase
tiktoken
pydantic
